app/scanner/openvas.py
"""OpenVAS/GVM API integration for network vulnerability scanning."""

from typing import Any


def run_openvas(
    host: str,
    api_key: str,
    targets: list[str],
    port: int = 9390,
) -> dict[str, Any]:
    """
    Trigger OpenVAS scan via API.
    Requires OpenVAS/GVM server running and python-gvm.
    targets: list of subnet CIDRs or host IPs
    """
    try:
        from gvm.protocols.http.openvasd import OpenvasdHttpAPIv1
    except ImportError:
        return {"status": "n/a", "message": "python-gvm not installed"}

    try:
        api = OpenvasdHttpAPIv1(host_name=host, port=port, api_key=api_key)
        # Create target and task, start scan
        # Implementation depends on OpenVAS API version
        return {"status": "info", "message": "OpenVAS integration - manual setup required", "targets": targets}
    except Exception as e:
        return {"status": "error", "message": str(e)}
