requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
paramiko>=3.4.0
Jinja2>=3.1.0
weasyprint>=60.0
python-multipart>=0.0.6
python-gvm>=23.0.0
